[doc] Skip migrate documentation for Python 3+
The migrate plugin is Python-2-only, and so should be its documentation.

Tiny caveat: this will leave red links within other documentation, but
removing references is difficult and probably not worth the effort.

A broken reference should alert the user that the functionality might
not be available well enough on its own.

Closes: #358
Approved by: Conan-Kudo